Change to Capability type code
