+++ src/listener/listener.js
@@ -14,7 +14,7 @@ import time from "./time.js";
 import frequency from "./frequency.js";
 import attribute from "./attribute.js";
 import capability from "./capability.js";
-import capability_type from "./capability_type.js";
+import capability_type from "./CapabilityTypeListener.js";
 import data_type from "./data_type.js";
 import requirement from "./requirement.js";
 import operation_def from "./operation_def.js";
